Integrated contacts list to structure of project
