--- app/src/main/java/com/github/factotum_sdp/factotum/contacts_list/ContactsFragment.kt
@@ -1,32 +1,0 @@
-package com.github.factotum_sdp.factotum.contacts_list
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.github.factotum_sdp.factotum.R
-
-class ContactsFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_contacts, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        val recycler = view.findViewById <RecyclerView>(R.id.contacts_recycler_view) // connect the recycler view to the layout
-        //the recycler is just the way we chose to represent the list of contacts
-        recycler.apply {
-            layoutManager = LinearLayoutManager(context)    //recycler needs a layout manager and an adapter which we set here
-            adapter = ContactsRecyclerAdapter()
-        }
-    }
-}
