--- app/src/main/java/com/github/factotum_sdp/factotum/contacts_list/DiaryMainActivity.kt
@@ -1,13 +1,0 @@
-package com.github.factotum_sdp.factotum.contacts_list
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.github.factotum_sdp.factotum.R
-
-class DiaryMainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main_diary)
-    }
-}
