+++ app/src/main/res/layout/fragment_contacts.xml
@@ -2,7 +2,10 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:orientation="horizontal">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="horizontal"
+    tools:context=".ui.directory.DirectoryFragment"
+    android:id="@+id/fragment_directory_directors_parent">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
