--- app/src/main/res/navigation/contacts_nav_graph.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/contactsNavGraph"
-    app:startDestination="@id/contactsFragment">
-
-    <fragment
-        android:id="@+id/contactsFragment"
-        android:name="com.github.factotum_sdp.factotum.contacts_list.ContactsFragment"
-        android:label="ContactsFragment" >
-        <action
-            android:id="@+id/action_contactsFragment_to_contactDetailsFragment"
-            app:destination="@id/contactDetailsFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/contactDetailsFragment"
-        android:name="com.github.factotum_sdp.factotum.contacts_list.ContactDetailsFragment"
-        android:label="ContactDetailsFragment" >
-        <argument
-            android:name="position"
-            app:argType="integer"
-            android:defaultValue="0" />
-        <action
-            android:id="@+id/action_contactDetailsFragment_to_contactsFragment"
-            app:destination="@id/contactsFragment" />
-    </fragment>
-</navigation>
