+++ app/src/main/res/navigation/nav_graph.xml
@@ -19,10 +19,22 @@
         android:id="@+id/directoryFragment"
         android:name="com.github.factotum_sdp.factotum.ui.directory.DirectoryFragment"
         android:label="fragment_directory"
-        tools:layout="@layout/fragment_directory" />
+        tools:layout="@layout/fragment_directory" >
+        <action
+            android:id="@+id/action_directoryFragment_to_contactDetailsFragment2"
+            app:destination="@id/contactDetailsFragment2" />
+    </fragment>
     <fragment
         android:id="@+id/mapsFragment"
         android:name="com.github.factotum_sdp.factotum.ui.maps.MapsFragment"
         android:label="fragment_maps"
         tools:layout="@layout/fragment_maps" />
+    <fragment
+        android:id="@+id/contactDetailsFragment2"
+        android:name="com.github.factotum_sdp.factotum.ui.directory.ContactDetailsFragment"
+        android:label="ContactDetailsFragment" >
+        <action
+            android:id="@+id/action_contactDetailsFragment2_to_directoryFragment"
+            app:destination="@id/directoryFragment" />
+    </fragment>
 </navigation>
